+++ src/entities/Enemy_Constrainer.hpp
@@ -23,8 +23,7 @@ public:
         dirX = -1;
       else if (stats.pos.x <= 0)
         dirX = 1;
-      stats.pos.y -=
-          std::sinf(actorCooldown) * stats.speed * p.dt * (float)dirY;
+      stats.pos.y -= std::sin(actorCooldown) * stats.speed * p.dt * (float)dirY;
       if (stats.pos.y >= 530.0f)
         dirY = -1;
       else if (stats.pos.y <= 0)
