+++ src/projectiles/Proj_Paren.hpp
@@ -10,12 +10,12 @@ public:
 
   void update(float dt, float bulletTime) override {
     if (bulletTime <= 2.0f)
-      stats.size += std::sinf(bulletTime) * 0.8f;
+      stats.size += std::sin(bulletTime) * 0.8f;
     else
-      stats.size -= std::sinf(bulletTime) * 0.8f;
+      stats.size -= std::sin(bulletTime) * 0.8f;
 
-    stats.pos.x += std::cosf(bulletTime) * stats.vel.x * dt;
-    stats.pos.y += std::cosf(bulletTime) * stats.vel.y * dt;
+    stats.pos.x += std::cos(bulletTime) * stats.vel.x * dt;
+    stats.pos.y += std::cos(bulletTime) * stats.vel.y * dt;
     stats.lifetime -= dt;
   };
